refactor: extract template string code generation into entries/ modules
Extract 5 entry template generators from index.ts and app-dev-server.ts
into packages/vinext/src/entries/:

- app-rsc-entry.ts: generateRscEntry() + AppRouterConfig interface
- app-ssr-entry.ts: generateSsrEntry()
- app-browser-entry.ts: generateBrowserEntry()
- pages-server-entry.ts: generateServerEntry() (closure -> explicit params)
- pages-client-entry.ts: generateClientEntry() (closure -> explicit params)

app-dev-server.ts is now a thin re-export barrel (12 lines, was 3034).
index.ts reduced by ~1238 lines (2507, was 3745).

Snapshot tests added in tests/entry-templates.test.ts to guard against
template drift during future refactors.

Closes #253
